Create issue_template.md
Added a new markdown file called "issue_template.md" in the new directory "docs" with some basic but useful suggestions for issue formatting. This include 3 headings:

What needs to be done - the basic description and lists of tasks that need to be done for the issue to be complete.

What doing this will achieve - why the issue needs to be completed for the project

Next Steps - how the new additions will be used, or how they relate to other future issues.